requirements.txt
discord.py
requests
yt_dlp